Add Function Water In Box, Add Boss Data, Add Boss Room

+++ Assets/Minh/Buff_Effect/BuffEffectManager/BuffEffectFactory.cs
@@ -2,7 +2,7 @@
 
 public static class BuffEffectFactory
 {
-    public static IBuffEffect CreateEffect(string effectName, float value, float duration)
+    public static IBuffEffect CreateEffect(string effectName, int value, int duration)
     {
         switch (effectName.ToLower())
         {
